Move a class in views.py into views directory package
