+++ kuhub/views/__init__.py
@@ -1,0 +1,14 @@
+from .gened_type_view import GenEdTypeListView
+from .group_view import GroupView, create_group, join
+from .notification_view import NotificationView
+from .post_feature_view import (
+    create_post, post_detail, edit_post, report_post, like_post, dislike_post
+)
+from .profile_view import (
+    profile_view, profile_settings, toggle_follow,
+    followers_page, following_page
+)
+from .review_view import ReviewHubView
+from .subject_detail_view import SubjectDetailView
+from .summary_view import SummaryHubView
+from .tricks_view import TricksHubView
