+++ kuhub/views/review_view.py
@@ -1,0 +1,34 @@
+from django.db.models import QuerySet
+from django.views import generic
+from kuhub.filters import PostFilter
+from kuhub.models import Post, Profile
+
+
+class ReviewHubView(generic.ListView):
+    """Redirect to Review-Hub page for review posts."""
+
+    queryset = Post.objects.all().filter(tag_id=1).order_by('-post_date')
+    template_name: str = 'kuhub/review.html'
+    context_object_name: str = 'posts_list'
+
+    def get_queryset(self) -> QuerySet[Post]:
+        queryset = super().get_queryset()
+        self.filterset = PostFilter(self.request.GET, queryset=queryset)
+        return self.filterset.qs
+
+    def get_context_data(self, **kwargs):
+        """Add like and dislike icon styles to context."""
+        context = super().get_context_data(**kwargs)
+
+        profiles_list = [Profile.objects.filter(user=post.username).first()
+                         for post in context['posts_list']]
+
+        context['like_icon_styles'] = [post.like_icon_style(self.request.user)
+                                       for post in context['posts_list']]
+        context['dislike_icon_styles'] = [
+            post.dislike_icon_style(self.request.user) for post in
+            context['posts_list']]
+        context['profiles_list'] = profiles_list
+        context['form'] = self.filterset.form
+
+        return context
