+++ kuhub/views/summary_view.py
@@ -1,0 +1,37 @@
+from django.db.models import QuerySet
+from django.views import generic
+from kuhub.filters import PostDownloadFilter
+from kuhub.models import PostDownload, Profile
+
+
+class SummaryHubView(generic.ListView):
+    """Redirect to Summary-Hub page for summary posts."""
+
+    queryset = ((PostDownload.objects
+                .select_related('post_id__tag_id'))
+                .order_by('-post_id__post_date').all())
+    template_name: str = 'kuhub/summary.html'
+    context_object_name: str = 'summary_post_list'
+
+    def get_queryset(self) -> QuerySet[PostDownload]:
+        """Return PostDownload objects with tag_id=2 and order by post_date."""
+        queryset = super().get_queryset()
+        self.filterset = PostDownloadFilter(self.request.GET, queryset=queryset)
+        return self.filterset.qs
+
+    def get_context_data(self, **kwargs):
+        """Add like and dislike icon styles to context."""
+        context = super().get_context_data(**kwargs)
+
+        profiles_list = [Profile.objects.filter(user=post.post_id.username).first()
+                         for post in context['summary_post_list']]
+
+        context['like_icon_styles'] = [post.like_icon_style(self.request.user)
+                                       for post in context['summary_post_list']]
+        context['dislike_icon_styles'] = [
+            post.dislike_icon_style(self.request.user) for post in
+            context['summary_post_list']]
+        context['profiles_list'] = profiles_list
+        context['form'] = self.filterset.form
+
+        return context
